+++ lib/qonfig/commands/definition/load_from_self.rb
@@ -2,7 +2,7 @@
 
 # @api private
 # @since 0.2.0
-class Qonfig::Commands::LoadFromSelf < Qonfig::Commands::Base
+class Qonfig::Commands::Definition::LoadFromSelf < Qonfig::Commands::Base
   # @return [String, Symbol]
   #
   # @api private