+++ lib/qonfig/commands/definition/load_from_yaml.rb
@@ -2,7 +2,7 @@
 
 # @api private
 # @since 0.2.0
-class Qonfig::Commands::LoadFromYAML < Qonfig::Commands::Base
+class Qonfig::Commands::Definition::LoadFromYAML < Qonfig::Commands::Base
   # @return [String]
   #
   # @api private