+++ lib/qonfig/data_set.rb
@@ -13,19 +13,24 @@ class Qonfig::DataSet # rubocop:disable Metrics/ClassLength
   extend Qonfig::Validator::DSL
 
   class << self
+    # @param base_dataset_klass [Class<Qonfig::DataSet>]
     # @param config_klass_definitions [Proc]
     # @return [Qonfig::DataSet]
     #
     # @api public
     # @since 0.16.0
-    def build(&config_klass_definitions)
-      Class.new(self, &config_klass_definitions).new
+    def build(base_dataset_klass = self, &config_klass_definitions)
+      unless base_dataset_klass <= Qonfig::DataSet
+        raise(Qonfig::ArgumentError, 'Base class should be a type of Qonfig::DataSet')
+      end
+
+      Class.new(base_dataset_klass, &config_klass_definitions).new
     end
   end
 
   # @return [Qonfig::Settings]
   #
-  # @api private
+  # @api public
   # @since 0.1.0
   attr_reader :settings
 
@@ -67,6 +72,65 @@ class Qonfig::DataSet # rubocop:disable Metrics/ClassLength
     thread_safe_definition do
       raise Qonfig::FrozenSettingsError, 'Frozen config can not be reloaded' if frozen?
       load!(settings_map, &configurations)
+    end
+  end
+
+  # @param file_path [String, Symbol]
+  # @option format [String, Symbol]
+  # @option strict [Boolean]
+  # @option expose [NilClass, String, Symbol] Environment key
+  # @return [void]
+  #
+  # @see Qonfig::DataSet#load_setting_values_from_file
+  #
+  # @api public
+  # @since 0.17.0
+  def load_from_file(file_path, format: :dynamic, strict: true, expose: nil)
+    thread_safe_access do
+      load_setting_values_from_file(file_path, format: format, strict: strict, expose: expose)
+    end
+  end
+
+  # @param file_path [String]
+  # @option strict [Boolean]
+  # @option expose [NilClass, String, Symbol] Environment key
+  # @return [void]
+  #
+  # @see Qonfig::DataSet#load_from_file
+  #
+  # @api public
+  # @since 0.17.0
+  def load_from_yaml(file_path, strict: true, expose: nil)
+    load_from_file(file_path, format: :yml, strict: strict, expose: expose)
+  end
+
+  # @param file_path [String]
+  # @option strict [Boolean]
+  # @option expose [NilClass, String, Symbol] Environment key
+  # @return [void]
+  #
+  # @see Qonfig::DataSet#load_from_file
+  #
+  # @api public
+  # @since 0.17.0
+  def load_from_json(file_path, strict: true, expose: nil)
+    load_from_file(file_path, format: :json, strict: strict, expose: expose)
+  end
+
+  # @option format [String, Symbol]
+  # @option strict [Boolean]
+  # @option expose [NilClass, String, Symbol]
+  # @return [void]
+  #
+  # @api public
+  # @since 0.17.0
+  def load_from_self(format: :dynamic, strict: true, expose: nil)
+    caller_location = caller(1, 1).first
+
+    thread_safe_access do
+      load_setting_values_from_file(
+        :self, format: format, strict: strict, expose: expose, caller_location: caller_location
+      )
     end
   end
 
@@ -184,6 +248,17 @@ class Qonfig::DataSet # rubocop:disable Metrics/ClassLength
     thread_safe_access { settings.__subset__(*keys) }
   end
 
+  # @param key_path [Array<String, Symbol>]
+  # @return [Boolean]
+  #
+  # @api public
+  # @since 0.17.0
+  def key?(*key_path)
+    thread_safe_access { settings.__has_key__(*key_path) }
+  end
+  alias_method :option?, :key?
+  alias_method :setting?, :key?
+
   # @return [void]
   #
   # @api public
@@ -278,6 +353,40 @@ class Qonfig::DataSet # rubocop:disable Metrics/ClassLength
     end
   end
 
+  # @param temporary_configurations [Hash<Symbol|String,Any>]
+  # @param arbitary_code [Block]
+  # @return [void]
+  #
+  # @api public
+  # @since 0.17.0
+  def with(temporary_configurations = {}, &arbitary_code)
+    with_arbitary_access do
+      begin
+        original_settings = @settings
+
+        temporary_settings = self.class.build.dup.tap do |copied_config|
+          copied_config.configure(temporary_configurations)
+        end.settings
+
+        @settings = temporary_settings
+        yield if block_given?
+      ensure
+        @settings = original_settings
+      end
+    end
+  end
+
+  # @return [Qonfig::DataSet]
+  #
+  # @api public
+  # @since 0.17.0
+  def dup
+    thread_safe_definition do
+      self.class.build.tap do |duplicate|
+        duplicate.configure(to_h)
+      end
+    end
+  end
 
   private
 
@@ -315,6 +424,16 @@ class Qonfig::DataSet # rubocop:disable Metrics/ClassLength
     yield(settings) if block_given?
   end
 
+  # @return [void]
+  #
+  # @api private
+  # @since 0.17.0
+  def call_instance_management_commands
+    self.class.instance_commands.each do |instance_command|
+      instance_command.call(self, settings)
+    end
+  end
+
   # @param settings_map [Hash]
   # @param configurations [Proc]
   # @return [void]
@@ -324,7 +443,31 @@ class Qonfig::DataSet # rubocop:disable Metrics/ClassLength
   def load!(settings_map = {}, &configurations)
     build_validator
     build_settings
+    call_instance_management_commands
     apply_settings(settings_map, &configurations)
+  end
+
+  # @param file_path [String, Symbol]
+  # @option format [String, Symbol]
+  # @option strict [Boolean]
+  # @option expose [NilClass, String, Symbol]
+  # @option callcer_location [NilClass, String]
+  # @return [void]
+  #
+  # @see Qonfig::Commands::Instantiation::ValuesFile
+  #
+  # @api private
+  # @since 0.17.0
+  def load_setting_values_from_file(
+    file_path,
+    format: :dynamic,
+    strict: true,
+    expose: nil,
+    caller_location: nil
+  )
+    Qonfig::Commands::Instantiation::ValuesFile.new(
+      file_path, caller_location, format: format, strict: strict, expose: expose
+    ).call(self, settings)
   end
 
   # @param instructions [Proc]
@@ -343,5 +486,9 @@ class Qonfig::DataSet # rubocop:disable Metrics/ClassLength
   # @since 0.2.0
   def thread_safe_definition(&instructions)
     @__lock__.thread_safe_definition(&instructions)
+  end
+
+  def with_arbitary_access(&instructions)
+    @__lock__.with_arbitary_access(&instructions)
   end
 end