+++ lib/qonfig/dsl.rb
@@ -10,12 +10,16 @@ module Qonfig::DSL
     # @api private
     # @since 0.1.0
     def extended(child_klass)
-      child_klass.instance_variable_set(:@commands, Qonfig::CommandSet.new)
+      child_klass.instance_variable_set(:@definition_commands, Qonfig::CommandSet.new)
+      child_klass.instance_variable_set(:@instance_commands, Qonfig::CommandSet.new)
 
       child_klass.singleton_class.prepend(Module.new do
         def inherited(child_klass)
-          child_klass.instance_variable_set(:@commands, Qonfig::CommandSet.new)
-          child_klass.commands.concat(commands)
+          child_klass.instance_variable_set(:@definition_commands, Qonfig::CommandSet.new)
+          child_klass.instance_variable_set(:@instance_commands, Qonfig::CommandSet.new)
+
+          child_klass.definition_commands.concat(definition_commands)
+          child_klass.instance_commands.concat(instance_commands)
           super
         end
       end)
@@ -25,9 +29,17 @@ module Qonfig::DSL
   # @return [Qonfig::CommandSet]
   #
   # @api private
-  # @since 0.1.0
-  def commands
-    @commands
+  # @since 0.17.0
+  def definition_commands
+    @definition_commands
+  end
+
+  # @return [Qonfig::CommandSet]
+  #
+  # @api private
+  # @since 0.17.0
+  def instance_commands
+    @instance_commands
   end
 
   # @param key [Symbol, String]
@@ -42,9 +54,9 @@ module Qonfig::DSL
   # @since 0.1.0
   def setting(key, initial_value = nil, &nested_settings)
     if block_given?
-      commands << Qonfig::Commands::AddNestedOption.new(key, nested_settings)
+      definition_commands << Qonfig::Commands::Definition::AddNestedOption.new(key, nested_settings)
     else
-      commands << Qonfig::Commands::AddOption.new(key, initial_value)
+      definition_commands << Qonfig::Commands::Definition::AddOption.new(key, initial_value)
     end
   end
 
@@ -56,7 +68,7 @@ module Qonfig::DSL
   # @api private
   # @sine 0.1.0
   def compose(data_set_klass)
-    commands << Qonfig::Commands::Compose.new(data_set_klass)
+    definition_commands << Qonfig::Commands::Definition::Compose.new(data_set_klass)
   end
 
   # @param file_path [String]
@@ -68,7 +80,9 @@ module Qonfig::DSL
   # @api public
   # @since 0.2.0
   def load_from_yaml(file_path, strict: true)
-    commands << Qonfig::Commands::LoadFromYAML.new(file_path, strict: strict)
+    definition_commands << Qonfig::Commands::Definition::LoadFromYAML.new(
+      file_path, strict: strict
+    )
   end
 
   # @return [void]
@@ -79,7 +93,10 @@ module Qonfig::DSL
   # @since 0.2.0
   def load_from_self(format: :yaml)
     caller_location = caller(1, 1).first
-    commands << Qonfig::Commands::LoadFromSelf.new(caller_location, format: format)
+
+    definition_commands << Qonfig::Commands::Definition::LoadFromSelf.new(
+      caller_location, format: format
+    )
   end
 
   # @option convert_values [Boolean]
@@ -91,7 +108,7 @@ module Qonfig::DSL
   # @api public
   # @since 0.2.0
   def load_from_env(convert_values: false, prefix: nil, trim_prefix: false)
-    commands << Qonfig::Commands::LoadFromENV.new(
+    definition_commands << Qonfig::Commands::Definition::LoadFromENV.new(
       convert_values: convert_values,
       prefix: prefix,
       trim_prefix: trim_prefix
@@ -105,7 +122,7 @@ module Qonfig::DSL
   # @api public
   # @since 0.5.0
   def load_from_json(file_path, strict: true)
-    commands << Qonfig::Commands::LoadFromJSON.new(file_path, strict: strict)
+    definition_commands << Qonfig::Commands::Definition::LoadFromJSON.new(file_path, strict: strict)
   end
 
   # @param file_path [String]
@@ -117,7 +134,9 @@ module Qonfig::DSL
   # @api public
   # @since 0.7.0
   def expose_yaml(file_path, strict: true, via:, env:)
-    commands << Qonfig::Commands::ExposeYAML.new(file_path, strict: strict, via: via, env: env)
+    definition_commands << Qonfig::Commands::Definition::ExposeYAML.new(
+      file_path, strict: strict, via: via, env: env
+    )
   end
 
   # @param file_path [String]
@@ -129,7 +148,9 @@ module Qonfig::DSL
   # @api public
   # @since 0.14.0
   def expose_json(file_path, strict: true, via:, env:)
-    commands << Qonfig::Commands::ExposeJSON.new(file_path, strict: strict, via: via, env: env)
+    definition_commands << Qonfig::Commands::Definition::ExposeJSON.new(
+      file_path, strict: strict, via: via, env: env
+    )
   end
 
   # @option env [Symbol, String]
@@ -141,6 +162,25 @@ module Qonfig::DSL
   # @since 0.14.0
   def expose_self(env:, format: :yaml)
     caller_location = caller(1, 1).first
-    commands << Qonfig::Commands::ExposeSelf.new(caller_location, env: env, format: format)
+
+    definition_commands << Qonfig::Commands::Definition::ExposeSelf.new(
+      caller_location, env: env, format: format
+    )
+  end
+
+  # @param file_path [String]
+  # @option format [String, Symbol]
+  # @option strict [Boolean]
+  # @option expose [NilClass, String, Symbol] Environment key
+  # @return [void]
+  #
+  # @api public
+  # @since 0.17.0
+  def values_file(file_path, format: :dynamic, strict: false, expose: nil)
+    caller_location = caller(1, 1).first
+
+    instance_commands << Qonfig::Commands::Instantiation::ValuesFile.new(
+      file_path, caller_location, format: format, strict: strict, expose: expose
+    )
   end
 end