+++ lib/qonfig/loaders/yaml.rb
@@ -7,10 +7,23 @@ class Qonfig::Loaders::YAML < Qonfig::Loaders::Basic
     # @param data [String]
     # @return [Object]
     #
+    # @raise [Qonfig::YAMLLoaderParseError]
+    #
     # @api private
     # @since 0.2.0
     def load(data)
       ::YAML.load(ERB.new(data).result)
+    rescue ::Psych::SyntaxError => error
+      raise(
+        Qonfig::YAMLLoaderParseError.new(
+          error.file,
+          error.line,
+          error.column,
+          error.offset,
+          error.problem,
+          error.context
+        ).tap { |exception| exception.set_backtrace(error.backtrace) }
+      )
     end
 
     # @return [Object]