+++ lib/qonfig/settings.rb
@@ -1,5 +1,8 @@
 # frozen_string_literal: true
- # rubocop:disable Metrics/ClassLength, Layout/ClassStructure
+
+# @api private
+# @since 0.1.0
+# rubocop:disable Metrics/ClassLength, Layout/ClassStructure
 class Qonfig::Settings # NOTE: Layout/ClassStructure is disabled only for CORE_METHODS constant
   require_relative 'settings/callbacks'
   require_relative 'settings/lock'
@@ -286,6 +289,15 @@ class Qonfig::Settings # NOTE: Layout/ClassStructure is disabled only for CORE_M
     value.is_a?(Qonfig::Settings)
   end
 
+  # @param key_path [Array<Symbol, String>]
+  # @return [Boolean]
+  #
+  # @api private
+  # @since 0.17.0
+  def __has_key__(*key_path)
+    __lock__.thread_safe_access { __is_key_exists__(*key_path) }
+  end
+
   private
 
   # @return [Qonfig::Settings::Lock]
@@ -337,6 +349,18 @@ class Qonfig::Settings # NOTE: Layout/ClassStructure is disabled only for CORE_M
   # @since 0.17.0
   def __root_setting_keys__
     __options__.keys
+  end
+
+  # @param key_path [Array<String, Symbol>]
+  # @return [Boolean]
+  #
+  # @api private
+  # @since 0.17.0
+  def __is_key_exists__(*key_path)
+    __deep_access__(*key_path)
+    true
+  rescue Qonfig::UnknownSettingError
+    false
   end
 
   # @param block [Proc]